更新:model/User.age,entity/UserTable.age long型からint型へ変更
影響:HomeController,FinishController,User登録関連,前述したクラスの各テストクラス

+++ src/main/java/com/example/demo/controller/HomeController.java
@@ -34,7 +34,7 @@ public class HomeController {
 			@Value("${addressapi.origin}") String addressAPIOrigin,
 			Model model) {
 		
-		int newUserNo = userService.getNewNo();
+		long newUserNo = userService.getNewNo();
 		user.setNo(newUserNo);
 		
 		model.addAttribute("addressAPIOrigin", addressAPIOrigin);
