+++ src/test/java/com/example/demo/entity/UserTableTest.java
@@ -22,7 +22,7 @@ public class UserTableTest extends BaseEntityTest {
 	/** 氏名 */
 	private String initName;
 	/** 年齢 */
-	private long initAge;
+	private int initAge;
 	/** 誕生日 */
 	private Date initBirthday;
 	/** 比較用インスタンス */
@@ -32,7 +32,7 @@ public class UserTableTest extends BaseEntityTest {
 	public void initUserTable() throws ParseException {
 		initNo		 = 10L;
 		initName	 = "テストユーザ";
-		initAge		 = 30L;
+		initAge		 = 30;
 		initBirthday = new Date(new SimpleDateFormat("yyyy/MM/dd").parse("1995/01/01").getTime());
 		
 		ut = new UserTable();
@@ -52,7 +52,7 @@ public class UserTableTest extends BaseEntityTest {
 		public void _010101_getter_and_setter() throws ParseException, IOException {
 			long	expectedNo		 = 10L;
 			String	expectedName	 = "テストユーザ";
-			long	expectedAge		 = 30L;
+			int	expectedAge		 = 30;
 			Date	expectedBirthday = new Date(new SimpleDateFormat("yyyy/MM/dd").parse("1995/01/01").getTime());
 			
 			UserTable ut = new UserTable();
@@ -127,7 +127,7 @@ public class UserTableTest extends BaseEntityTest {
 			UserTable notSameAge = new UserTable();
 			notSameAge.setNo(initNo);
 			notSameAge.setName(initName);
-			notSameAge.setAge(initAge + 1L);
+			notSameAge.setAge(initAge + 1);
 			notSameAge.setBirthday(initBirthday);
 			
 			assertThat(ut.equals(notSameAge)).isFalse();
